Add files via upload
It was my first a temped to making a cacheing system to cppcheck externally.   I wrote it a couple of years ago.  I don't even remember the state of the code or if it is working at this point.  I just wanted a place to put it before I lost it.